+++ packages/stateful/hooks/index.ts
@@ -13,6 +13,7 @@ export * from './useInboxApi'
 export * from './useInstantiateAndExecute'
 export * from './useLoadingDaoCardInfos'
 export * from './useLoadingGovProposal'
+export * from './useLoadingGovProposalWalletVoteInfo'
 export * from './useMembership'
 export * from './useProposalPolytoneState'
 export * from './useProposalVetoState'
